Migrated From Shuffleboard to Elastic & Glass
It appears that Shuffleboard no longer works when trying to publish data to the robot (it can still receive data though). I have tried both simulation and actual code on multiple laptops, and other teams have had similar problems: https://www.chiefdelphi.com/t/shuffleboard-inputs-no-longer-working/482144
Therefore, we are migrating from Shuffleboard to a combination of Elastic and Glass. Both are installed automatically with the WPILib installer.
Elastic will be used for a Driver dashboard.
Glass will be used for debugging (since it is more reliable and easier to use than SmartDashboard).
There's also a default layout for elastic placed inside of the <deploy> folder for the robot.