Removed addRequirement a bunch
While this is supposed to be used to literally tell the code to dedicate the subsystem to that command and is generally very needed, this is also the line of code that caused us to do nothing at Week 0.
When we're adding commands to the AutoChooser, we instantiate each command. If we have addRequirement in the constructor of the command, then it can break things if any part of our bot is disabled (commented out in RobotMap).